refactor: TrackInfo
