+++ src/components/player/index.ts
@@ -1,4 +1,3 @@
 export * from './PlayerControls';
 export * from './PlayPauseButton';
 export * from './Progress';
-export * from './TrackInfo/TrackInfo';
